requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
matplotlib
opencv-python
streamlit
tqdm
Pillow
